Optimized fast rotation.
Added various optimizations to the fast rotations without hurting
quality too much.

Some of the changes include:
- Baking normalization into the transform.
- Specialized implementation of length 64 and length 256 transforms.
- Restricting the max length of the block transforms to 256.
- Switching to float32 instead of float64 (did not improve performance
  on my machine, but should generally be more efficient and amenable to
  later SIMD optimization, as well as using less memory).
- Performing the random swaps with greater locality.
- Storing the random signs directly as float32 values to avoid casts.

Overall this reduces encoding time by around 40%.

+++ adapters/repos/db/vector/compressionhelpers/compression.go
@@ -673,7 +673,7 @@ func RestoreRQCompressor(
 	outputDim int,
 	rounds int,
 	swaps [][]Swap,
-	signs [][]int8,
+	signs [][]float32,
 	store *lsmkv.Store,
 	allocChecker memwatch.AllocChecker,
 ) (VectorCompressor, error) {
@@ -744,7 +744,7 @@ func RestoreRQMultiCompressor(
 	outputDim int,
 	rounds int,
 	swaps [][]Swap,
-	signs [][]int8,
+	signs [][]float32,
 	store *lsmkv.Store,
 	allocChecker memwatch.AllocChecker,
 ) (VectorCompressor, error) {
