+++ adapters/repos/db/vector/hnsw/deserializer_test.go
@@ -611,7 +611,7 @@ func TestDeserializerTotalReadRQ(t *testing.T) {
 				{I: 17, J: 19},
 			},
 		},
-		Signs: [][]int8{
+		Signs: [][]float32{
 			{1, -1, 1, -1},
 			{1, -1, 1, -1},
 			{1, -1, 1, -1},
